+++ Exploration/DirectEmitter.php
@@ -13,9 +13,6 @@ use Heptacom\HeptaConnect\Portal\Base\Emission\Contract\EmitterStackInterface;
 
 final class DirectEmitter extends EmitterContract
 {
-    /**
-     * @psalm-var DatasetEntityCollection<DatasetEntityContract>
-     */
     private DatasetEntityCollection $entities;
 
     public function __construct(
@@ -36,9 +33,6 @@ final class DirectEmitter extends EmitterContract
         yield from $this->emitNext($stack, $externalIds, $context);
     }
 
-    /**
-     * @psalm-return DatasetEntityCollection<DatasetEntityContract>
-     */
     public function getEntities(): DatasetEntityCollection
     {
         return $this->entities;