+++ Job/Transition/Contract/EmittedEntitiesToJobsConverterInterface.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace Heptacom\HeptaConnect\Core\Job\Transition\Contract;
 
 use Heptacom\HeptaConnect\Core\Job\JobCollection;
-use Heptacom\HeptaConnect\Dataset\Base\Contract\DatasetEntityContract;
 use Heptacom\HeptaConnect\Dataset\Base\DatasetEntityCollection;
 use Heptacom\HeptaConnect\Portal\Base\StorageKey\Contract\PortalNodeKeyInterface;
 
@@ -13,8 +12,6 @@ interface EmittedEntitiesToJobsConverterInterface
 {
     /**
      * Converts freshly emitted entities into jobs like receive jobs.
-     *
-     * @param DatasetEntityCollection<DatasetEntityContract> $entities
      */
     public function convert(PortalNodeKeyInterface $portalNodeKey, DatasetEntityCollection $entities): JobCollection;
 }