+++ Portal/AbstractPortalNodeContext.php
@@ -11,6 +11,7 @@ use Heptacom\HeptaConnect\Portal\Base\Portal\Contract\PortalNodeContextInterface
 use Heptacom\HeptaConnect\Portal\Base\Portal\Contract\PortalStorageInterface;
 use Heptacom\HeptaConnect\Portal\Base\StorageKey\Contract\PortalNodeKeyInterface;
 use Psr\Container\ContainerInterface;
+use Psr\Log\LoggerInterface;
 
 abstract class AbstractPortalNodeContext implements PortalNodeContextInterface
 {
@@ -43,6 +44,11 @@ abstract class AbstractPortalNodeContext implements PortalNodeContextInterface
     public function getStorage(): PortalStorageInterface
     {
         return $this->containerFacade->getStorage();
+    }
+
+    public function getLogger(): LoggerInterface
+    {
+        return $this->containerFacade->getLogger();
     }
 
     public function getContainer(): ContainerInterface