+++ Portal/Contract/PortalRegistryInterface.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace Heptacom\HeptaConnect\Core\Portal\Contract;
 
-use Heptacom\HeptaConnect\Portal\Base\Exploration\Contract\ExplorerContract;
 use Heptacom\HeptaConnect\Portal\Base\Portal\Contract\PortalContract;
+use Heptacom\HeptaConnect\Portal\Base\Portal\Contract\PortalExtensionContract;
 use Heptacom\HeptaConnect\Portal\Base\Portal\PortalExtensionCollection;
 use Heptacom\HeptaConnect\Portal\Base\StorageKey\Contract\PortalNodeKeyInterface;
 
@@ -20,7 +20,7 @@ interface PortalRegistryInterface
     public function getPortal(PortalNodeKeyInterface $portalNodeKey): PortalContract;
 
     /**
-     * Get @see ExplorerContract instances that can be used on the portal node and are set to be active.
+     * Get @see PortalExtensionContract instances that can be used on the portal node and are set to be active.
      */
     public function getPortalExtensions(PortalNodeKeyInterface $portalNodeKey): PortalExtensionCollection;
 }