+++ Reception/PostProcessing/MarkAsFailedPostProcessor.php
@@ -13,13 +13,11 @@ use Heptacom\HeptaConnect\Storage\Base\Action\IdentityError\Create\IdentityError
 use Heptacom\HeptaConnect\Storage\Base\Action\IdentityError\Create\IdentityErrorCreatePayloads;
 use Heptacom\HeptaConnect\Storage\Base\Contract\Action\IdentityError\IdentityErrorCreateActionInterface;
 use Heptacom\HeptaConnect\Storage\Base\PrimaryKeySharingMappingStruct;
-use Psr\Log\LoggerInterface;
 
 final class MarkAsFailedPostProcessor extends PostProcessorContract
 {
     public function __construct(
-        private IdentityErrorCreateActionInterface $identityErrorCreateAction,
-        private LoggerInterface $logger
+        private IdentityErrorCreateActionInterface $identityErrorCreateAction
     ) {
     }
 
@@ -29,7 +27,6 @@ final class MarkAsFailedPostProcessor extends PostProcessorContract
             $event->getContext()->getPostProcessingBag()->of(MarkAsFailedData::class),
             static fn (MarkAsFailedData $data) => $data
         );
-        $logger = $event->getContext()->getContainer()->get(LoggerInterface::class) ?? $this->logger;
 
         /** @var MarkAsFailedData $data */
         foreach ($markAsFailedData as $data) {
@@ -52,7 +49,7 @@ final class MarkAsFailedPostProcessor extends PostProcessorContract
                 }
             }
 
-            $logger->error(LogMessage::MARK_AS_FAILED_ENTITY_IS_UNMAPPED(), [
+            $event->getContext()->getLogger()->error(LogMessage::MARK_AS_FAILED_ENTITY_IS_UNMAPPED(), [
                 'throwable' => $data->getThrowable(),
                 'data' => $data,
                 'code' => 1637456198,